requirements.txt
fastapi[standard]
uvicorn
python-dotenv
google-adk
soundfile
openai-whisper
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
mcp>=1.0.0